requirements.txt
Jinja2==3.1.*
Mastodon.py==1.8.*
scipy==1.13.*
beautifulsoup4==4.12.*
requests==2.32.*
